+++ Echo/src/utopia/echo/model/request/chat/BufferedOrStreamedChatRequest.scala
@@ -6,9 +6,7 @@ import utopia.annex.util.ResponseParseExtensions._
 import utopia.disciple.http.response.ResponseParser
 import utopia.echo.controller.EchoContext
 import utopia.echo.controller.parser.StreamedReplyMessageResponseParser
-import utopia.echo.model.ChatMessage
 import utopia.echo.model.response.chat.{BufferedReplyMessage, StreamedOrBufferedReplyMessage}
-import utopia.flow.collection.immutable.Empty
 import utopia.flow.parse.json.JsonParser
 import utopia.flow.util.logging.Logger
 
@@ -20,11 +18,9 @@ import scala.concurrent.{ExecutionContext, Future}
   * @author Mikko Hilpinen
   * @since 21.07.2024, v1.0
   */
-class BufferedOrStreamedChat(override val message: ChatMessage,
-                             override val conversationHistory: Seq[ChatMessage] = Empty,
-                             override val stream: Boolean = false, testDeprecation: => Boolean = false)
-                            (implicit exc: ExecutionContext, jsonParser: JsonParser, log: Logger)
-	extends Chat[StreamedOrBufferedReplyMessage]
+case class BufferedOrStreamedChatRequest(params: ChatParams, stream: Boolean = false)
+                                        (implicit exc: ExecutionContext, jsonParser: JsonParser, log: Logger)
+	extends ChatRequest[StreamedOrBufferedReplyMessage]
 {
 	// ATTRIBUTES   --------------------------
 	
@@ -39,8 +35,6 @@ class BufferedOrStreamedChat(override val message: ChatMessage,
 	
 	
 	// IMPLEMENTED  --------------------------
-	
-	override def deprecated: Boolean = testDeprecation
 	
 	override def send(prepared: ApiClient.PreparedRequest): Future[RequestResult[StreamedOrBufferedReplyMessage]] =
 		prepared.send(responseParser)